requirements.txt
# Web / API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
streamlit==1.28.0
python-dotenv==1.0.0

# CLI / query language
groq

# Authentication
python-jose==3.5.0
PyJWT==2.8.0
cryptography==46.0.2
passlib==1.7.4
bcrypt>=4.0.1

# Database / migrations
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Pydantic
pydantic>=2.7.4,<3.0.0
pydantic-settings==2.1.0

# AI/ML & Vector DB
sentence-transformers==2.2.2
chromadb==0.4.15
huggingface-hub==0.19.4
openai-whisper==20231117

# LangChain (optional)
langchain==0.1.20
langchain-community>=0.0.38

# Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
pypdf2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2
pillow==10.1.0
python-magic==0.4.27

# Utilities
numpy==1.24.3
pandas==2.1.4
requests==2.31.0
tqdm==4.66.1

# Dev / Test
pytest==7.4.3
black==23.11.0